Compute accession register when elimnation action occurs Correct bug 4146

+++ sources/access/access-internal/access-internal-core/src/main/java/fr/gouv/vitam/access/internal/core/AccessInternalModuleImpl.java
@@ -415,7 +415,7 @@ public class AccessInternalModuleImpl implements AccessInternalModule {
         String mimetype = null;
         String filename = null;
         String objectId = null;
-        Integer size = null;
+        Long size = null;
         if (finalversionsResponse != null) {
             if (finalversionsResponse.getFormatIdentification() != null &&
                 !finalversionsResponse.getFormatIdentification().getMimeType().isEmpty()) {
